requirements.txt
google_api_python_client==2.70.0
google_auth_oauthlib==0.4.6
matplotlib==3.6.2
Pillow==9.3.0
protobuf==4.21.12
